Refactor builder to module
