+++ src/lava/magma/runtime/message_infrastructure/multiprocessing.py
@@ -4,7 +4,7 @@
 import typing as ty
 if ty.TYPE_CHECKING:
     from lava.magma.core.process.process import AbstractProcess
-    from lava.magma.compiler.builder import PyProcessBuilder, \
+    from lava.magma.compiler.builders.builder import PyProcessBuilder, \
         AbstractRuntimeServiceBuilder
 
 import multiprocessing as mp
